+++ hoka/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 5.1.1 - 2020-01-14
+
+### Changed
+
+  - MultipartBodyParser optimization (getBoundaryIndex).
+
 ## 5.1.0 - 2019-03-13
 
 ### Changed
@@ -68,6 +74,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Initial  revision.
   
 ---  
-_Copyright 2017-2019 MicroEJ Corp. All rights reserved._  
+_Copyright 2017-2020 MicroEJ Corp. All rights reserved._  
 _This library is provided in source code for use, modification and test, subject to license terms._  
 _Any modification of the source code will break MicroEJ Corp. warranties on the whole library._